Play with words(tolower and toupper)
A computer program receives a word and a character from user. All characters of the word is turned lowercase using int tolower(int c) function from <ctype.h> library. The character is checked if it is alphabetic using int isalpha(int c) function from <ctype.h> library. If it is not alphabetic, a new character input is requested from user until it is alphabetic. If it is alphabetic, the character is turned lowercase, just in case. The character is searched through the word and turned uppercase at every position it is found. Final state of the word is displayed.